discovered branches
The last 5 weeks from the latest update on April 7th, I was working on the app continously, but hidden. Now, I decided to code more in public in a separate branch.
The work is about "making it right" after the "make it work" prototype currently on the web. The classes and structures are now modeled properly.